move dedup methods to mod.rs file
